Assignment-3/requirements.txt
numpy==1.26.4
scikit-learn==1.2.2
pillow==10.2.0
pandas==2.1.4
cvxopt=1.3.2